feat: Update Lucide React import in developer profile page
This commit updates the import statement for the Lucide React library in the developer profile page. It adds the "MapPin" icon to the list of imported icons. This change allows the user's location to be displayed with an icon in the profile page, improving the visual representation of the user's information.